+++ packages/docs/gatsby-ssr.tsx
@@ -1,8 +1,6 @@
-const path = require('path');
 const proc = require('process');
-const { readFileSync } = require('fs');
 
-const dangerousHTML = (html) => {
+const makeDangerousHTML = (html) => {
   return { __html: html };
 };
 
@@ -10,35 +8,17 @@ const HeadComponentsJs = (content) => (
   <script type="text/javascript" dangerouslySetInnerHTML={content}></script>
 );
 
-const specialPages = ['/404.html', '/404/', '/blog/', '/contact/', '/', '/search/'];
+exports.onRenderBody = ({ setPreBodyComponents }) => {
+  const env = proc.env.NODE_ENV == 'production' ? 'prod' : 'dev';
 
-exports.onRenderBody = ({ pathname, setPreBodyComponents }) => {
-  const env = proc.env.NODE_ENV;
-  const siteSection = pathname;
-  let pageName;
-  if (!specialPages.includes(siteSection)) {
-    const file = siteSection.slice(0, siteSection.length - 1);
-    const fp = path.join('content', `${file}.mdx`);
-    const frontMatterRegExp = new RegExp('(?<=-{3})(.*)(?=-{3})', 'sg');
-    const titleRegExp = new RegExp('(title:){1}(.*)', 'g');
-    try {
-      const fileContent = readFileSync(path.resolve(fp)).toString();
-      const [frontmatter] = fileContent.toString().match(frontMatterRegExp);
-      pageName = [...frontmatter.matchAll(titleRegExp)][0][2].trim();
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
+  const tealiumSnippet = `
+    (function(a,b,c,d){
+        a='https://tealium-tags.cms.gov/cms-design/${env}/utag.js';
+        b=document;c='script';d=b.createElement(c);d.src=a;d.type='text/javascript';d.async=true;
+        a=b.getElementsByTagName(c)[0];a.parentNode.insertBefore(d,a);
+    })();`;
 
-  const utag_data = `var utag_data = {
-    content_language: 'en',
-    content_type: 'html',
-    logged_in: 'false',
-    page_name: '${pageName}',
-    page_type: 'false',
-    site_environment: '${env}',
-    site_section: '${siteSection}',
-  }`;
+  const dangerousTealiumSnippet = makeDangerousHTML(tealiumSnippet);
 
-  setPreBodyComponents(HeadComponentsJs(dangerousHTML(utag_data)));
+  setPreBodyComponents(HeadComponentsJs(dangerousTealiumSnippet));
 };
