--- packages/docs/ssr-helper.js
@@ -1,11 +1,0 @@
-var utag_data = {
-  content_language: 'en',
-  content_type: 'html',
-  logged_in: 'false',
-  page_name: 'TEST PAGE',
-  page_type: 'false',
-  site_environment: 'dev',
-  site_section: 'no value',
-};
-
-console.log(utag_data);
